+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ready` option for `--type` flag in `account create` and `account import` commands (Argent wallet has rebranded as Ready)
 
+#### Changed
+
+- Braavos accounts with all class hashes are now supported 
+
 #### Deprecated
 
 - `argent` option for `--type` flag in `account create` and `account import` commands. Use `ready` instead